requirements.txt
streamlit
pdfplumber
python-docx
pandas
numpy
scikit-learn
fuzzywuzzy
sentence-transformers
